+++ mne/event.py
@@ -381,7 +381,7 @@ def find_stim_steps(raw, pad_start=None, pad_stop=None, merge=0,
     """
 
     # pull stim channel from config if necessary
-    stim_channel = _get_stim_channel(stim_channel)
+    stim_channel = _get_stim_channel(stim_channel, raw.info)
 
     picks = pick_channels(raw.info['ch_names'], include=stim_channel)
     if len(picks) == 0:
@@ -478,8 +478,9 @@ def find_events(raw, stim_channel=None, verbose=None, output='onset',
         Name of the stim channel or all the stim channels
         affected by the trigger. If None, the config variables
         'MNE_STIM_CHANNEL', 'MNE_STIM_CHANNEL_1', 'MNE_STIM_CHANNEL_2',
-        etc. are read. If these are not found, it will default to
-        'STI 014'.
+        etc. are read. If these are not found, it will fall back to
+        'STI 014' if present, then fall back to the first channel of type
+        'stim', if present.
     verbose : bool, str, int, or None
         If not None, override default verbose level (see mne.verbose).
     output : 'onset' | 'offset' | 'step'
@@ -579,7 +580,7 @@ def find_events(raw, stim_channel=None, verbose=None, output='onset',
     min_samples = min_duration * raw.info['sfreq']
 
     # pull stim channel from config if necessary
-    stim_channel = _get_stim_channel(stim_channel)
+    stim_channel = _get_stim_channel(stim_channel, raw.info)
 
     pick = pick_channels(raw.info['ch_names'], include=stim_channel)
     if len(pick) == 0: